# webapps/swagger.yaml
swagger: '2.0'
info:
  description: "This website for testing API of Test server project ISS. If any question can read more detail on http://swagger.io or on irc.freenode.net, #swagger."
  version: "1.0.0"
  title: "Swagger ISS"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "issadmin@celestica.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
consumes:
  - application/json
produces:
  - application/json
basePath: "/api"
...
paths:
...
  /settings:
    get: 
      tags: [Settings]
      operationId: controllers.setting_controller.get_settings
      summary: Get all settings
      parameters: 
        - name: limit
          in: query
          type: integer
          minimum: 0
          default: 100
      responses:
        200:
          description: Return setting
          schema:
            type: array
            items:
              $ref: '#/definitions/Setting'
    put:  
      tags: [Settings]
      operationId: controllers.setting_controller.put_setting
      summary: Create or update a setting
      parameters: 
        - name: setting
          in: body
          schema:
            $ref: '#/definitions/Setting'
      responses:
        200:
          description: Setting updated
        201:
          description: New setting created
  /settings/{setting_id}:
    get:  
      tags: [Settings]
      operationId: controllers.setting_controller.get_setting
      summary: Get a single setting
      parameters:
        - $ref: '#/parameters/setting_id'
      responses:
        200:
          description: Return pet
          schema:
            $ref: '#/definitions/Setting'
        404:
          description: Setting does not exist
    delete:  
      tags: [Settings]
      operationId: controllers.setting_controller.delete_setting
      summary: Remove a setting
      parameters:
        - $ref: '#/parameters/setting_id'
      responses:
        204:
          description: Setting was deleted
        404:
          description: Setting does not exist
...
parameters:
...
  setting_id:
    name: setting_id
    description: Setting's Unique identifier
    in: path
    type: integer
    required: true 
...
definitions:
...
  Setting:
    type: object
    required:
      - name 
    properties:
      id:
        type: integer
        description: Unique identifier
        example: 123 
      name:
        type: string
        description: Setting's name
        example: "IP Address"
        minLength: 1
        maxLength: 100
      value:
        type: string
        description: Setting's value
        example: "0.0.0.0"
        minLength: 0 
      created:
        type: string
        format: date-time
        description: Creation time
        example: "2015-07-07T15:49:51.230+02:00" 
...